requirements.txt
Django==3.0.7
djangorestframework_jwt==1.11.0
djangorestframework==3.11.0
